chore: adjust key logic to share information
