+++ specs/1106-KEYS-wallet_and_keys.md
@@ -2,7 +2,7 @@
 
 As a browser wallet user I want to view my vega wallet (and key pair(s)) So that I can see that I've been successful creating the wallet / see my key ID
 
-- I can see a list of the keys in my wallet (<a name="1106-KEYS-001" href="#1106-KEYS-001">1106-KEYS-001</a>)
+- I can see a list of the keys in my wallet with [associated info](./1126-VKEY-vega_key.md) (<a name="1106-KEYS-001" href="#1106-KEYS-001">1106-KEYS-001</a>)
 - I can copy the public key ID to my clipboard (<a name="1106-KEYS-002" href="#1106-KEYS-002">1106-KEYS-002</a>)
 - I can see information of where to go to deposit and manage my assets (<a name="1106-KEYS-003" href="#1106-KEYS-003">1106-KEYS-003</a>)
 - I can see where I am in the app when viewing my wallet and key pair(s) (<a name="1106-KEYS-004" href="#1106-KEYS-004">1106-KEYS-004</a>)
